requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
pydantic==1.10.9
python-dotenv==1.0.0
transformers==4.30.2
sentencepiece==0.1.99
sacremoses==0.0.53
torch==2.0.1
langdetect==1.0.9
tqdm==4.65.0 